gpgme: fix build
avoid this issue:

| rmdir: failed to remove `/srv/home/nitin/builds/build-gcc47/tmp/wERROR: Function failed: do_install (see /srv/home/nitin/builds/build-gcc47/tmp/work/i586-poky-linux/gpgme-1.3.1-r1/temp/log.do_install.36763 for further information)
| ork/i586-poky-linux/gpgme-1.3.1-r1/image/usr/libexec': No such file or directory
NOTE: package gpgme-1.3.1-r1: task do_install: Failed

no PR bump as no change in the output

+++ meta/recipes-support/gpgme/gpgme_1.3.1.bb
@@ -28,5 +28,7 @@ FILES_${PN} = "${libdir}/libgpgme.so.*"
 FILES_${PN}-dev += "${bindir}/gpgme-config ${datadir}/common-lisp/source/gpgme/*"
 
 do_install_append() {
-	rmdir ${D}/usr/libexec
+	if [ -d ${D}/usr/libexec ] ; then
+		rmdir ${D}/usr/libexec
+	fi
 }
